Move validate/parsed.js into lib
