+++ index.js
@@ -1,4 +1,4 @@
 exports.format = require('./lib/format');
 exports.migrate = require('./lib/migrate');
 exports.validate = require('./lib/validate');
-exports.validate.parsed = require('./validate/parsed');
+exports.validate.parsed = require('./lib/validate/parsed');
