+++ lib/validate.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var jsonlint = require('jsonlint-lines-primitives');
-var validate = require('../validate/parsed');
+var validate = require('./validate/parsed');
 
 /**
  * Validate a Mapbox GL Style given as a string of JSON. Returns an array
