+++ package.json
@@ -8,8 +8,8 @@
     "gl-style-format": "bin/gl-style-format"
   },
   "scripts": {
-    "test": "jshint lib/*.js migrations/*.js validate/*.js && tape test/*.js test/migrations/*.js",
-    "doc": "cat lib/*.js validate/*.js | dox --raw --skipSingleStar | doxme > API.md",
+    "test": "jshint lib/*.js lib/*/*.js migrations/*.js && tape test/*.js test/migrations/*.js",
+    "doc": "cat lib/*.js lib/*/*.js | dox --raw --skipSingleStar | doxme > API.md",
     "cov": "istanbul cover ./node_modules/.bin/tape test/*.js test/migrations/*.js && coveralls < ./coverage/lcov.info"
   },
   "repository": {
